+++ packages/terra-core-docs/src/terra-dev-site/test/form-textarea/AutoResizableTextarea.test.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Textarea from '../../../Textarea';
+import Textarea from 'terra-form-textarea';
 import styles from './common/Textarea.test.module.scss';
 
 const cx = classNames.bind(styles);
 
 const textarea = () => (
   <div className={cx('content-wrapper')}>
-    <Textarea size="small" ariaLabel="label" />
+    <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" ariaLabel="label" />
   </div>
 );