+++ packages/terra-core-docs/src/terra-dev-site/test/form-textarea/AutoResizableTextareaMobileMaxTouchPoints.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Textarea from '../../../Textarea';
+import Textarea from 'terra-form-textarea';
 import styles from './common/Textarea.test.module.scss';
 
 const cx = classNames.bind(styles);